# optimize performance of  `rb_str_resurrect`