Thank you for applying the proposed approach in e2cabc2.

Unfortunately, there were issues in the commit (and other related commits) that failed to gain the maximum from the division of fast and slow passes.  This PR addresses the issue.

The numbers I see when running the code below is as follows:

compiler | trunk | this PR |
---|---:|---:|
gcc 4.8.2 (linux) | 1.31 | 1.27 |
gcc 5.1.0 (OS X) | 1.49 | 1.42 |
clang-600.0.56 (OS X) | 1.58 | 1.51 |

benchmark code:
```
require "benchmark"
N = 20_000_000
GC.start; GC.start # magic
Benchmark.bm{|bm|
  bm.report{N.times{str = "a"}}
}
```